requirements.txt
bcrypt==4.1.2
cffi==1.16.0
cryptography==42.0.5
matlabengine==9.12.21
numpy==1.24.4
pandas==2.0.3
paramiko==3.4.0
pycparser==2.21
PyNaCl==1.5.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
tzdata==2024.1
